--- src/Unima.Application/Commands/Project/CreateGitProject/CreateGitProjectCommand.cs
@@ -1,17 +1,0 @@
-﻿using MediatR;
-
-namespace Unima.Application.Commands.Project.CreateGitProject
-{
-    public class CreateGitProjectCommand : IRequest<bool>
-    {
-        public string RepositoryUrl { get; set; }
-
-        public string Branch { get; set; }
-
-        public string Username { get; set; }
-
-        public string Password { get; set; }
-
-        public string OutputPath { get; set; }
-    }
-}
