--- src/Unima.Application/Commands/Project/CreateGitProject/CreateGitProjectCommandHandler.cs
@@ -1,14 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
-
-namespace Unima.Application.Commands.Project.CreateGitProject
-{
-    public class CreateGitProjectCommandHandler : IRequestHandler<CreateGitProjectCommand, bool>
-    {
-        public Task<bool> Handle(CreateGitProjectCommand request, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(false);
-        }
-    }
-}
