--- src/Unima.Application/Commands/Project/CreateGitProject/CreateGitProjectCommandValidator.cs
@@ -1,8 +1,0 @@
-﻿using FluentValidation;
-
-namespace Unima.Application.Commands.Project.CreateGitProject
-{
-    public class CreateGitProjectCommandValidator : AbstractValidator<CreateGitProjectCommand>
-    {
-    }
-}
