+++ src/Unima.Application/Models/GitInfo.cs
@@ -1,0 +1,13 @@
+﻿namespace Unima.Application.Models
+{
+    public class GitInfo
+    {
+        public string Url { get; set; }
+
+        public string Branch { get; set; }
+
+        public string Username { get; set; }
+
+        public string Password { get; set; }
+    }
+}
