+++ src/Unima.Console/packages.config
@@ -9,6 +9,8 @@
   <package id="Costura.Fody" version="4.1.0" targetFramework="net472" />
   <package id="Fody" version="6.0.5" targetFramework="net461" developmentDependency="true" />
   <package id="Humanizer.Core" version="2.3.3" targetFramework="net461" />
+  <package id="LibGit2Sharp" version="0.26.1" targetFramework="net472" />
+  <package id="LibGit2Sharp.NativeBinaries" version="2.0.289" targetFramework="net472" />
   <package id="log4net" version="2.0.8" targetFramework="net461" />
   <package id="McMaster.Extensions.CommandLineUtils" version="2.2.4" targetFramework="net461" />
   <package id="MediatR" version="7.0.0" targetFramework="net472" />
