+++ src/Unima.Core/Git/IGitCloner.cs
@@ -1,0 +1,7 @@
+﻿namespace Unima.Core.Git
+{
+    public interface IGitCloner
+    {
+        void CloneProject(string repositoryUrl, string branch, string username, string password, string outputPath);
+    }
+}
