+++ src/Unima.Core/Unima.Core.csproj
@@ -64,13 +64,14 @@
     <Compile Include="Execution\Compilation\IProjectCompiler.cs" />
     <Compile Include="Execution\Report\Html\HtmlOnlyBodyReportCreator.cs" />
     <Compile Include="Execution\Report\Summary\TextSummaryReportCreator.cs" />
-    <Compile Include="Execution\Runners\ITestRunner.cs" />
     <Compile Include="Execution\MutationDocumentExecutor.cs" />
     <Compile Include="Execution\Result\TestResult.cs" />
     <Compile Include="Execution\Result\TestSuiteResult.cs" />
+    <Compile Include="Execution\Runners\ITestRunner.cs" />
     <Compile Include="Execution\Runners\ITestRunnerFactory.cs" />
     <Compile Include="Execution\Runners\ITestRunnerClient.cs" />
     <Compile Include="Execution\TestRunnerDependencyFilesHandler.cs" />
+    <Compile Include="Git\IGitCloner.cs" />
     <Compile Include="Location\MutationLocationInfo.cs" />
     <Compile Include="Loggers\AzureMutationRunLogger.cs" />
     <Compile Include="Loggers\IMutationRunLogger.cs" />
