+++ src/Unima.Git.Console/Commands/ICommand.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Unima.Git.Commands
+{
+    public interface ICommand
+    {
+        Task<int> RunAsync();
+    }
+}
