+++ src/Unima.Git.Console/Commands/RootCommand.cs
@@ -1,0 +1,21 @@
+﻿using System.Threading.Tasks;
+using McMaster.Extensions.CommandLineUtils;
+
+namespace Unima.Git.Commands
+{
+    public class RootCommand : ICommand
+    {
+        private CommandLineApplication _app;
+
+        public RootCommand(CommandLineApplication app)
+        {
+            _app = app;
+        }
+
+        public Task<int> RunAsync()
+        {
+            _app.ShowHelp();
+            return Task.FromResult(-1);
+        }
+    }
+}
