+++ src/Unima.Git.Console/packages.config
@@ -1,0 +1,35 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="CommandLine.Core.CommandLineUtils" version="1.1.0" targetFramework="net472" />
+  <package id="CommandLine.Core.Hosting" version="1.1.0" targetFramework="net472" />
+  <package id="CommandLine.Core.Hosting.Abstractions" version="1.1.0" targetFramework="net472" />
+  <package id="CommandLineUtils.Extensions" version="1.1.0" targetFramework="net472" />
+  <package id="Humanizer.Core" version="2.3.3" targetFramework="net472" />
+  <package id="LibGit2Sharp" version="0.26.1" targetFramework="net472" />
+  <package id="LibGit2Sharp.NativeBinaries" version="2.0.289" targetFramework="net472" />
+  <package id="McMaster.Extensions.CommandLineUtils" version="2.2.4" targetFramework="net472" />
+  <package id="MedallionShell" version="1.6.0" targetFramework="net472" />
+  <package id="Microsoft.DotNet.PlatformAbstractions" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration.Binder" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Configuration.EnvironmentVariables" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.DependencyInjection" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.DependencyModel" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.FileProviders.Abstractions" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.FileProviders.Physical" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.FileSystemGlobbing" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Localization.Abstractions" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Logging" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Logging.Abstractions" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Options" version="2.1.0" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Primitives" version="2.1.0" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net472" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net472" />
+  <package id="System.Memory" version="4.5.0" targetFramework="net472" />
+  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net472" />
+  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net472" />
+  <package id="System.ValueTuple" version="4.4.0" targetFramework="net472" />
+</packages>
