+++ src/Unima.Infrastructure/FodyWeavers.xml
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-  <Anotar.Log4Net />/&gt;
+  <Anotar.Log4Net />
 </Weavers>
