+++ src/Unima.TestRunner.Console/App.config
@@ -5,7 +5,6 @@
     </startup>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <probing privatePath="lib"/>
     </assemblyBinding>
   </runtime>
 </configuration>
