+++ src/Unima.TestRunner.Console/packages.config
@@ -1,4 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="MedallionShell" version="1.6.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net461" />
+  <package id="NUnit.Engine" version="3.9.0" targetFramework="net472" />
+  <package id="StyleCop.Analyzers" version="1.0.2" targetFramework="net472" developmentDependency="true" />
+  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net472" />
+  <package id="xunit.abstractions" version="2.0.3" targetFramework="net472" />
+  <package id="xunit.runner.utility" version="2.4.1" targetFramework="net472" />
 </packages>
