--- src/Unima.TestRunner/NUnit/NUnitTestRunFinishedTestEventListener.cs
@@ -1,26 +1,0 @@
-﻿using System.Threading;
-using NUnit.Engine;
-
-namespace Unima.TestRunner.NUnit
-{
-    public class NUnitTestRunFinishedTestEventListener : ITestEventListener
-    {
-        private readonly ManualResetEvent _manualResetEvent;
-
-        public NUnitTestRunFinishedTestEventListener(ManualResetEvent manualResetEvent)
-        {
-            _manualResetEvent = manualResetEvent;
-        }
-
-        public bool TestRunFinished { get; private set; }
-
-        public void OnTestEvent(string report)
-        {
-            if (report.Contains("test-run"))
-            {
-                TestRunFinished = true;
-                _manualResetEvent.Set();
-            }
-        }
-    }
-}
