--- src/Unima.TestRunner/packages.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Anotar.Log4Net.Fody" version="5.1.0" targetFramework="net461" />
-  <package id="Fody" version="6.0.5" targetFramework="net461" developmentDependency="true" />
-  <package id="log4net" version="2.0.8" targetFramework="net461" />
-  <package id="MedallionShell" version="1.6.0" targetFramework="net461" />
-  <package id="NUnit.Engine" version="3.9.0" targetFramework="net461" />
-  <package id="StyleCop.Analyzers" version="1.0.2" targetFramework="net461" developmentDependency="true" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net461" />
-  <package id="xunit.abstractions" version="2.0.3" targetFramework="net461" />
-  <package id="xunit.runner.utility" version="2.4.1" targetFramework="net461" />
-</packages>
