+++ src/Unima/Bootstrapper.cs
@@ -10,9 +10,11 @@ using Unima.Application.Extensions;
 using Unima.Application.Logs;
 using Unima.Core.Execution.Compilation;
 using Unima.Core.Execution.Runners;
+using Unima.Core.Git;
 using Unima.Helpers.Displayers;
 using Unima.Helpers.Openers.Tabs;
 using Unima.Infrastructure;
+using Unima.Infrastructure.Git;
 using Unima.Sections.Loading;
 using Unima.Sections.Shell;
 using Unima.TestRunner;
@@ -79,6 +81,7 @@ namespace Unima
             Container.RegisterType<IProjectCompiler, Compiler>();
             Container.RegisterType<IMutationDocumentCompiler, Compiler>();
             Container.RegisterType<ITestRunnerClient, TestRunnerConsoleClient>();
+            Container.RegisterType<IGitCloner, GitCloner>(new ContainerControlledLifetimeManager());
         }
     }
 }
